refactor: extract integration-settings routes and types into modules
Move integration types into types/integrations.ts and convert types.ts
to a types/ directory module. Extract integration-settings route handlers
into routes/integration-settings.ts with shared route primitives in
routes/utils.ts. This reduces router.ts by ~390 lines and establishes
the pattern for future route and type extractions.
